+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## [Unreleased]
 ### Added
 - Added logger classes generation for autowire by Fully Qualified Class Name.
+### Changed
+- Renamed `LoggerHandler` into `LoggerCollection`.
 
 ## [1.0.0] - 2019-04-14
 ### Added